GEODE-6779: adjustments based on code review
- Move diskStore presence checking to DiskStoreCommandsUtils
- Move tests to newly created DiskStoreCommandsUtilsTest

Authored-by: Joris Melchior <joris.melchior@gmail.com>
