+++ geode-core/src/test/java/org/apache/geode/management/internal/cli/commands/GfshCommandJUnitTest.java
@@ -14,31 +14,20 @@
  */
 package org.apache.geode.management.internal.cli.commands;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.spy;
 
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.management.ObjectName;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mockito;
 
-import org.apache.geode.management.DistributedSystemMXBean;
 import org.apache.geode.management.cli.GfshCommand;
 import org.apache.geode.management.internal.exceptions.EntityNotFoundException;
 
 public class GfshCommandJUnitTest {
 
-  private String memberName = "memberONe";
-  private String diskStoreName = "diskStoreOne";
   private GfshCommand command;
 
   @Before
@@ -66,59 +55,5 @@ public class GfshCommandJUnitTest {
     doReturn(Collections.emptySet()).when(command).findMembersIncludingLocators(members, null);
     assertThatThrownBy(() -> command.getMembersIncludingLocators(members, null))
         .isInstanceOf(EntityNotFoundException.class);
-  }
-
-  @Test
-  public void diskStoreBeanAndMemberBeanDiskStoreExists() throws Exception {
-    Map<String, String[]> memberDiskStore = new HashMap<>();
-    memberDiskStore.put(memberName, new String[] {diskStoreName});
-    ObjectName objectName = new ObjectName("");
-
-    DistributedSystemMXBean distributedSystemMXBean = Mockito.mock(DistributedSystemMXBean.class);
-    doReturn(memberDiskStore).when(distributedSystemMXBean).listMemberDiskstore();
-    doReturn(objectName).when(distributedSystemMXBean).fetchDiskStoreObjectName(any(), any());
-
-    assertThat(command.diskStoreBeanAndMemberBeanDiskStoreExists(distributedSystemMXBean,
-        memberName, diskStoreName)).isTrue();
-  }
-
-  @Test
-  public void diskStoreBeanExistsAndMemberDiskStoreNotFound() throws Exception {
-    Map<String, String[]> memberDiskStore = new HashMap<>();
-    memberDiskStore.put(memberName, new String[] {});
-    ObjectName objectName = new ObjectName("");
-
-    DistributedSystemMXBean distributedSystemMXBean = Mockito.mock(DistributedSystemMXBean.class);
-    doReturn(memberDiskStore).when(distributedSystemMXBean).listMemberDiskstore();
-    doReturn(objectName).when(distributedSystemMXBean).fetchDiskStoreObjectName(any(), any());
-
-    assertThat(command.diskStoreBeanAndMemberBeanDiskStoreExists(distributedSystemMXBean,
-        memberName, diskStoreName)).isFalse();
-  }
-
-  @Test
-  public void diskStoreBeanNotFoundAndMemberDiskStoreExists() throws Exception {
-    Map<String, String[]> memberDiskStore = new HashMap<>();
-    memberDiskStore.put(memberName, new String[] {diskStoreName});
-
-    DistributedSystemMXBean distributedSystemMXBean = Mockito.mock(DistributedSystemMXBean.class);
-    doReturn(memberDiskStore).when(distributedSystemMXBean).listMemberDiskstore();
-    doThrow(new Exception("not found")).when(distributedSystemMXBean)
-        .fetchDiskStoreObjectName(any(), any());
-
-    assertThat(command.diskStoreBeanAndMemberBeanDiskStoreExists(distributedSystemMXBean,
-        memberName, diskStoreName)).isFalse();
-  }
-
-  @Test
-  public void diskStoreBeanExistsMemberDiskStoreIsNull() throws Exception {
-    ObjectName objectName = new ObjectName("");
-
-    DistributedSystemMXBean distributedSystemMXBean = Mockito.mock(DistributedSystemMXBean.class);
-    doReturn(null).when(distributedSystemMXBean).listMemberDiskstore();
-    doReturn(objectName).when(distributedSystemMXBean).fetchDiskStoreObjectName(any(), any());
-
-    assertThat(command.diskStoreBeanAndMemberBeanDiskStoreExists(distributedSystemMXBean,
-        memberName, diskStoreName)).isFalse();
   }
 }
